pyproject.toml
[project]
name = "modelbench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.4",
    "openai>=1.68.0",
    "requests>=2.32.3",
    "yt-dlp>=2025.2.19",
]
